+++ crates/pdk-api/src/lib.rs
@@ -1,10 +1,12 @@
 mod common;
+mod context;
 mod extension;
 mod macros;
 mod prompts;
 mod toolchain;
 
 pub use common::*;
+pub use context::*;
 pub use extension::*;
 pub use moon_project::ProjectFragment;
 pub use moon_task::TaskFragment;
