+++ crates/workspace/Cargo.toml
@@ -15,6 +15,7 @@ moon_config = { path = "../config" }
 moon_env_var = { path = "../env-var" }
 moon_feature_flags = { path = "../feature-flags" }
 moon_hash = { path = "../hash" }
+moon_pdk_api = { path = "../pdk-api" }
 moon_project = { path = "../project" }
 moon_project_builder = { path = "../project-builder" }
 moon_project_constraints = { path = "../project-constraints" }
